Minor standardization pass on scraper tests
This change applies some standardization to test cases
for several recently added scrapers.
